CN VAF Plot

Create a genome-wide copy number plot for one sample and (optionally) display mutation VAF.

Usage

copy_number_vaf_plot(
  this_sample_id,
  just_segments = FALSE,
  coding_only = FALSE,
  one_chrom,
  genes_to_label,
  this_seq_type = "genome",
  from_flatfile = TRUE,
  use_augmented_maf = TRUE,
  add_chr_prefix = FALSE
)

Arguments

this_sample_id

The sample_id for the sample to plot.

just_segments

Specify whether only the segments will be plotted (instead of mutation VAF). Default is FALSE.

coding_only

Optional. Set to TRUE to restrict to plotting only coding mutations.

one_chrom

Subset plot to one chromosome.

genes_to_label

Optional. Provide a vector of genes to label (if mutated). Can only be used with coding_only (see above).

this_seq_type

Seq type for returned CN segments. One of "genome" (default) or "capture".

from_flatfile

If set to true the function will use flatfiles instead of the database.

use_augmented_maf

Boolean statement if to use augmented maf, default is TRUE.

add_chr_prefix

If TRUE, "chr" prefix will be added to chr column. Default is FALSE.

Value

Nothing

Details

This function takes a sample ID and internally calls [GAMBLR::assign_cn_to_ssm] to get copy number segments for plotting. This plot is visualizing mutation VAFs per default, this can be turned off with setting `just_segments` to TRUE. This only plots the segments. The user can also restrict the plotted segments to coding regions. To do so, set `coding_only= TRUE`, and then specify the genes of interest (coding regions) with the `genes_to_label` (vector of genes).

Examples

if (FALSE) { # \dontrun{
#build plot
copy_number_vaf_plot(this_sample_id = "DOHH-2")

#coding only
copy_number_vaf_plot(this_sample_id = "DOHH-2",
                     coding_only = TRUE)
} # }