Assign a colour palette to metadata columns automatically and consistently.

Assign a colour palette to metadata columns automatically and consistently.

Usage

map_metadata_to_colours(
  metadataColumns,
  these_samples_metadata,
  column_alias = list(),
  verbose = FALSE,
  annoAlpha = 1
)

Arguments

metadataColumns

Names of the metadata columns to assign colours for.

these_samples_metadata

Metadata for just the samples you need colours for.

column_alias

A list of column_names with values indicating what gambl colour scheme they should use (e.g. pos_neg, pathology, lymphgen).

verbose

Set to TRUE to enable verbose mode (debugging messages.

annoAlpha

Optional alpha to apply to annotation colours.

Value

A named list

Examples

library(GAMBLR.open)

#get metadata
all_meta = suppressMessages(get_gambl_metadata() %>%
  dplyr::filter(pathology %in% c("FL","DLBCL")))

#get colours
all_cols = map_metadata_to_colours(
     metadataColumns = c(
         "lymphgen",
         "pathology"),
     these_samples_metadata = all_meta,
     column_alias = list("nothing" = "FL")
)