sample-level SV/SSM/CN reports in PDF

Construct pdf with sample-level plots, using minimum of arguments

comp_report(
  this_sample_id,
  export_individual_plots = FALSE,
  out = "./",
  seg_data,
  seg_path = NULL,
  maf_data,
  maf_path = NULL,
  this_seq_type = "genome",
  projection = "grch37"
)

Arguments

this_sample_id

Sample ID to be plotted in report.

export_individual_plots

Boolean parameter, set to TRUE to export individual plots.

out

Path to output folder. The default is the working directory.

seg_data

Optional parameter with copy number df already loaded into R.

seg_path

Optional parameter with path to external cn file.

maf_data

Optional parameter with maf like df already loaded into R.

maf_path

Optional parameter with path to external maf like file.

this_seq_type

Seq type for returned CN segments. One of "genome" (default) or "capture".

projection

Specify the projection you want the returned plots to be in reference to. Possible values are "grch37" and "hg38". Default is grch37.

Value

Nothing.

Details

This function runs the complete collection of `fancy_x_plots` for a specific sample ID (`this_sample`), with default parameters. The generated plots are put together into a two-page PDF. In addition, it is also possible to export all individual plots. This can be done by setting `export_individual_plots` to TRUE. It is also possible to use an already loaded seg file instead of using the `this_sample_id` parameter, this is done with the `seg_data` and `maf_data` parameters. Similarly, you can also point this function to a local file on disk with the `seg_path` and `maf_path` parameters.

Examples

if (FALSE) {
# create a PDF report for one sample, as well as exporting all 
# individual plots.
comp_report(this_sample_id = "HTMCP-01-06-00422-01A-01D",
            out = "./",
            export_individual_plots = TRUE)
}