Gene Cloud Plot

Make a word cloud of gene names from a MAF file based on mutation frequency.

prettyGeneCloud(
  maf_df,
  these_genes,
  other_genes,
  these_genes_colour = "#B2DF8A",
  other_genes_colour = "#bc42f5",
  colour_index,
  return_genes = FALSE,
  wordcloud_version = 2,
  zoomout = 1,
  fontFamily = "Arial"
)

Arguments

maf_df

A MAF-format data frame containing the mutations you want to summarize in a gene word cloud

these_genes

An optional vector of gene symbols (defaults to all lymphoma genes)

other_genes

An optional second vector of gene symbols to include in your cloud in a second colour

these_genes_colour

Specify the hex code of a colour to use for the first set of genes

other_genes_colour

Specify another hex code of a colour to use for the second set of genes

colour_index

Optional named character vector with a name for each gene in these_genes and a colour as the value

return_genes

Optionally request a vector of the genes in the plot. Default is FALSE.

wordcloud_version

Specify whether you want to use the original wordcloud package or wordcloud2 (1 or 2)

zoomout

Increase or decrease to fit more or less of the words into view. Only works when wordcloud_version is 2. Default 1

Value

data frame with counts for each gene

Details

Create a wordcloud from an incoming MAF. Required parameter is `maf_df`. Optional parameters are `these_genes`, `other_genes`, `these_genes_colour`, `other_genes_colour` and `colour_index`. If no genes are supplied when calling the function, this function will default to all lymphoma genes.

Examples

#get all coding SSM directly from GAMBLR.data
maf = GAMBLR.data::sample_data$grch37$maf

#build wordcloud
prettyGeneCloud(maf_df = maf,
                wordcloud_version = 2,
                zoomout=0.3)